requirements.txt
selenium==3.8.1
pymongo==3.6.0
Logbook==1.1.0
python_etcd==0.4.5
toml==0.9.4
Flask==0.12.2
kafka_python==1.4.2
pendulum==1.4.2
redis==2.10.6
requests==2.18.4
orator==0.9.7
sponge
